config/requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.3.0
billiard==4.2.1
blinker==1.8.2
boto3==1.41.5
botocore==1.41.5
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver==2.24.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==45.0.5
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2.1
django-allauth==65.10.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-oauth-toolkit==3.0.1
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.15
filelock==3.18.0
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.5.4
MarkupSafe==3.0.1
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
pipenv==2023.7.11
platformdirs==4.3.6
prompt_toolkit==3.0.51
psycopg==3.2.6
psycopg2-binary==2.9.10
pycoingecko==3.2.0
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.1.0
requests==2.32.4
s3transfer==0.15.0
selenium==4.29.0
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.14.1
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.3
virtualenv-clone==0.5.7
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
whitenoise==6.9.0
wsproto==1.2.0
